Add tss_reads.py to extract a list of read names supporting a peak region
It needs the bam file and the tss annotation groups TSS peak region
file. It generates a bed like file that has the coordinates and read
names for the reads that contributed to the region.

I did duplicate some code blocks between the two programs to make it
easier to just copy the the files around, this might need to be turned
into a proper python package.